python/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
cloud-sql-python-connector==1.18.2
SQLAlchemy==2.0.41
pg8000==1.31.2
python-dotenv==1.0.1
google-cloud-secret-manager==2.16.4
google-auth>=2.28.1
google-auth-oauthlib>=1.2.0
flask>=3.0.0
gunicorn>=21.2.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
google-cloud-kms==2.24.2
google-cloud-dlp==3.21.0
google-cloud-logging==3.10.0
google-cloud-monitoring==2.19.0
google-cloud-securitycenter==1.31.0
